+++ examples/nuxt/druxt-site/nuxt.config.js
@@ -3,12 +3,10 @@ export default {
   target: 'static',
   generate: { routes: ['/'] },
   telemetry: true,
-  components: true,
   modules: ['druxt-site'],
   druxt: {
     baseUrl,
     entity: { components: { fields: false } },
-    router: { pages: false },
     site: { theme: 'umami' }
   },
 }