+++ packages/views/src/components/DruxtView.vue
@@ -27,6 +27,20 @@ export default {
    */
   props: {
     /**
+     * Views contextual filters.
+     *
+     * @type {array}
+     * @default []
+     *
+     * @example
+     * <DruxtView :arguments="[1, 2, 3]" display-id="block_1" view-id="articles_aside" />
+     */
+    arguments: {
+      type: Array,
+      default: () => [],
+    },
+
+    /**
      * The View Display ID.
      *
      * @type {string}
@@ -375,6 +389,13 @@ export default {
       // Pagination.
       if (this.model.page) {
         query.page = this.model.page
+      }
+
+      // Contextual filters.
+      if (this.arguments.length) {
+        for (const index in this.arguments) {
+          query[`views-argument[${index}]`] = this.arguments[index]
+        }
       }
 
       // Exposed filters.